Changes that allow parallelization as jobs on HPC
