+++ ptm_torrent/huggingface/__main__.py
@@ -3,6 +3,7 @@ from os import environ
 import ptm_torrent.huggingface.createSchema as createSchema
 import ptm_torrent.huggingface.downloadJSON as downloadJSON
 import ptm_torrent.huggingface.downloadRepos as downloadRepos
+import ptm_torrent.huggingface.splitRepos as splitRepos
 import ptm_torrent.huggingface.setupFileSystem as setupFS
 
 if __name__ == "__main__":
@@ -26,5 +27,5 @@ if __name__ == "__main__":
 
     setupFS.main()
     downloadJSON.main()
-    downloadRepos.main(shrinkage)
+    splitRepos.main(shrinkage)
     createSchema.main()
