+++ ptm_torrent/huggingface/downloadRepos.py
@@ -1,7 +1,7 @@
 from typing import List
 
-import pandas
-from pandas import DataFrame
+import argparse
+import pathlib
 from progress.bar import Bar
 from progress.spinner import Spinner
 
@@ -10,51 +10,22 @@ from ptm_torrent.utils.fileSystem import readJSON, testForFile
 from ptm_torrent.utils.git import cloneRepo
 
 
-def readJSONData(json: dict) -> List[str]:
-    data: List[str] = []
+def main(url: pathlib.Path) -> None:
 
-    with Spinner("Reading JSON data...") as spinner:
-        obj: dict
-        for obj in json:
-            data.append(f"https://huggingface.co/{obj['id']}")
-            spinner.next()
-
-    return data
-
-
-def shrinkDataFrame(df: DataFrame, shrinkage: float = 0.1) -> DataFrame:
-    if shrinkage >= 1:
-        return df
-
-    dfSize: int = len(df)
-    lastRow: int = int(dfSize * shrinkage)
-
-    return df.iloc[0:lastRow, :]
-
-
-def main(shrinkage: float = 0.1) -> None:
-    if testForFile(path=hf.huggingface_HubMetadataPath) == False:
-        return False
-
-    print(f"Loading {hf.huggingface_HubMetadataPath} into DataFrame...")
-    df: DataFrame = pandas.read_json(hf.huggingface_HubMetadataPath)
-
-    print(f"Sorting DataFrame rows by download...")
-    df.sort_values(by="downloads", ascending=False, inplace=True)
-    df.reset_index(inplace=True)
-
-    print(f"Reducing the size of the DataFrame by {shrinkage * 100}%...")
-    df = shrinkDataFrame(df, shrinkage)
-
-    urls: List[str] = [f"https://huggingface.co/{id}" for id in df["id"]]
+    with open(url, "r") as f:
+        urls = f.readlines()
     with Bar(
         f"Cloning git repos to {hf.huggingface_ReposPath}...", max=len(urls)
     ) as bar:
         url: str
         for url in urls:
-            cloneRepo(url=url, rootGitClonePath=hf.huggingface_ReposPath)
+            cloneRepo(url=url.strip("\n"), rootGitClonePath=hf.huggingface_ReposPath)
             bar.next()
 
 
 if __name__ == "__main__":
-    main()
+
+    parser = argparse.ArgumentParser()
+    parser.add_argument("url_file", type=pathlib.Path)
+    args = parser.parse_args()
+    main(args.url_file)
